[FIX] mail: prevent deletion of attachment in email template
Step to reproduce:
1. Install `sale_management`
2. Open any email template (e.g., Sale: Order Confirmation).
3. Add an attachment to it
4. Create a Sale Order, confirm it, and click "Send by Email".
5. In the mail composer, remove the template attachment

Issue:
- The removed attachment is deleted from the database (`ir.attachment`).
  Consequently, the attachment is permanently removed from the source Email Template
  and will not appear in future emails.

Cause:
- The `onFileRemove` function in `MailComposerAttachmentList` calls the `unlink` method
  of the `attachmentUploadService` for every file removed, without considering the existing template attachment.

Solution:
1. Update code to include `res_model` in `Attachment`
2. In `onFileRemove`, check the `res_model` of the attachment.
3. If the `res_model` is not "mail.compose.message", skip the database deletion (unlink)
   and only remove it from the composer view.

opw-5163679

X-original-commit: 431d08373098248118d5e798ed551fe782ab754f

+++ addons/mail/static/src/core/web/mail_composer_attachment_list.js
@@ -20,7 +20,9 @@ export class MailComposerAttachmentList extends Many2ManyBinaryField {
     async onFileRemove(fileId) {
         super.onFileRemove(fileId);
         const attachment = this.mailStore["ir.attachment"].insert(fileId);
-        await this.attachmentUploadService.unlink(attachment);
+        if (attachment && attachment.res_model === "mail.compose.message") {
+            await this.attachmentUploadService.unlink(attachment);
+        }
         this.env.fullComposerBus.trigger("ATTACHMENT_REMOVED", {
             id: attachment.id,
         });
